% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peek.R
\name{peek}
\alias{peek}
\title{Take a peek at an object and then return it.}
\usage{
peek(.data, view = TRUE, print = FALSE)
}
\arguments{
\item{.data}{An R object that you want to see. Usually a matrix-like object,
such as a dataframe, matrix, or list.}

\item{view}{Whether to invoke a spreadsheet-style data viewer on `.data`.}

\item{print}{Whether to print `.data`.}
}
\value{
`.data`, after having viewed or printed it.
}
\description{
Take a peek at an object and then return it.
}
