% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peek.R
\name{pprint}
\alias{pprint}
\title{#' View an object and then return it.
#'
#' @param .data Any R object.
#'
#' @return \code{.data}, after having viewed it.
#' @examples
#' library(dplyr)
#' mtcars %>%
#'   filter(carb != 8) %>%
#'   select(carb, mpg, cyl, hp, drat) %>%
#'   group_by(carb) %>%
#'   summarise(min_mpg = min(mpg)) %>%
#'   pview() %>% # View the current state, then pass it to the next command.
#'   mutate(min_mpg_100 = min_mpg*100)
#'
#' @export
pview <- function(.data) {
  RStudioView <- as.environment("package:utils")$View
  RStudioView(.data)
  return(.data)
}
Print an object and then return it.}
\usage{
pprint(.data)
}
\arguments{
\item{.data}{Any R object.}
}
\value{
\code{.data}, after having printed it.
}
\description{
#' View an object and then return it.
#'
#' @param .data Any R object.
#'
#' @return \code{.data}, after having viewed it.
#' @examples
#' library(dplyr)
#' mtcars %>%
#'   filter(carb != 8) %>%
#'   select(carb, mpg, cyl, hp, drat) %>%
#'   group_by(carb) %>%
#'   summarise(min_mpg = min(mpg)) %>%
#'   pview() %>% # View the current state, then pass it to the next command.
#'   mutate(min_mpg_100 = min_mpg*100)
#'
#' @export
pview <- function(.data) {
  RStudioView <- as.environment("package:utils")$View
  RStudioView(.data)
  return(.data)
}
Print an object and then return it.
}
\examples{
library(dplyr)
mtcars \%>\%
  filter(carb != 8) \%>\%
  select(carb, mpg, cyl, hp, drat) \%>\%
  group_by(carb) \%>\%
  summarise(min_mpg = min(mpg)) \%>\%
  pprint() \%>\% # Print the current state, then pass it to the next command.
  mutate(min_mpg_100 = min_mpg*100)

}
