% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peek.R
\name{pview}
\alias{pview}
\title{View an object and then return it.}
\usage{
pview(.data)
}
\arguments{
\item{.data}{Any R object.}
}
\value{
\code{.data}, after having viewed it.
}
\description{
View an object and then return it.
}
